Split SignInModal and ForgotPasswordModal
