+++ content/src/ui/Modal.ForgotPassword.json
@@ -1,0 +1,41 @@
+{
+  "en": {
+    "title": "Forgot password",
+    "email": "E-mail",
+    "sendToEmail": "Send to my email",
+    "confirmSent": "Ok, got it!",
+    "checkEmailHeader": "Please check your email",
+    "emailSentInfo": "We sent you a link to reset your password.",
+    "errors": {
+      "auth/invalid-email": "Invalid email address",
+      "auth/network-request-failed": "Network request failed",
+      "unknown": "Unknown error"
+    }
+  },
+  "pt": {
+    "title": "Entrar",
+    "email": "Email",
+    "password": "Password",
+    "signIn": "Entrar",
+    "errors": {
+      "auth/invalid-email": "Email e/ou password errada",
+      "auth/wrong-password": "Email e/ou password errada",
+      "auth/user-not-found": "Email não existe",
+      "auth/user-disabled": "Email inativo"
+    }
+  },
+  "sv": {
+    "title": "Glömt lösenord",
+    "email": "Email",
+    "sendToEmail": "Skicka till min email",
+    "confirmSent": "Ok, jag förstår!",
+    "checkEmailHeader": "Kolla din epost",
+    "emailSentInfo": "Vi har skickat en länk för att återställa lösenord.",
+    "errors": {
+      "auth/invalid-email": "Ogiltig email",
+      "auth/network-request-failed": "Nätverksbegäran misslyckades",
+      "unknown": "Okänt fel"
+    }
+  },
+  "es": {}
+}
