+++ content/src/ui/Modal.SignIn.json
@@ -1,26 +1,16 @@
 {
   "en": {
     "title": "Sign in",
-    "resetTitle": "Forgot password",
     "email": "E-mail",
     "password": "Password",
     "signIn": "Sign in",
     "reset": "Forgot Password?",
-    "sendToEmail": "Send to my email",
-    "confirmSent": "Ok, got it!",
-    "checkEmailHeader": "Please check your email",
-    "emailSentInfo": "We sent you a link to reset your password.",
     "errors": {
       "auth/invalid-email": "Wrong e-mail and/or password",
       "auth/wrong-password": "Wrong e-mail and/or password",
       "auth/user-not-found": "Wrong e-mail and/or password",
       "auth/user-disabled": "Wrong e-mail and/or password",
       "auth/network-request-failed": "Network request failed"
-    },
-    "resetErrors": {
-      "auth/invalid-email": "Invalid email address",
-      "auth/network-request-failed": "Network request failed",
-      "unknown": "Unknown error"
     }
   },
   "pt": {
@@ -37,25 +27,15 @@
   },
   "sv": {
     "title": "Logga in",
-    "resetTitle": "Glömt lösenord",
     "email": "Email",
     "password": "Lösenord",
     "signIn": "Logga in",
     "reset": "Glömt Lösenord?",
-    "sendToEmail": "Skicka till min email",
-    "confirmSent": "Ok, jag förstår!",
-    "checkEmailHeader": "Kolla din epost",
-    "emailSentInfo": "Vi har skickat en länk för att återställa lösenord.",
     "errors": {
       "auth/invalid-email": "Fel email och/eller lösenord",
       "auth/wrong-password": "Fel email och/eller lösenord",
       "auth/user-not-found": "Fel email och/eller lösenord",
       "auth/user-disabled": "Fel email och/eller lösenord",
-      "auth/network-request-failed": "Nätverksbegäran misslyckades",
-      "unknown": "Okänt fel"
-    },
-    "resetErrors": {
-      "auth/invalid-email": "Ogiltig email",
       "auth/network-request-failed": "Nätverksbegäran misslyckades",
       "unknown": "Okänt fel"
     }
